+++ compiler-tests/src/test/data/diagnostic/bindingcontainer/IllegalTypes.kt
@@ -19,6 +19,9 @@ sealed class SealedClass
 <!BINDING_CONTAINER_ERROR!>@BindingContainer<!>
 sealed interface SealedInterface
 
+<!BINDING_CONTAINER_ERROR!>@BindingContainer<!>
+open class OpenClass
+
 class Containing {
   <!BINDING_CONTAINER_ERROR!>@BindingContainer<!>
   inner class InnerClass