+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -811,6 +811,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
     }
 
     @Test
+    @TestMetadata("IncludedClassBindingContainer.kt")
+    public void testIncludedClassBindingContainer() {
+      run("IncludedClassBindingContainer.kt");
+    }
+
+    @Test
     @TestMetadata("MultibindsOnlyInContainer.kt")
     public void testMultibindsOnlyInContainer() {
       run("MultibindsOnlyInContainer.kt");