+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsContributionProvidersBoxTestGenerated.java
@@ -549,6 +549,12 @@ public class JsContributionProvidersBoxTestGenerated extends AbstractJsContribut
     }
 
     @Test
+    @TestMetadata("IncludedClassBindingContainer.kt")
+    public void testIncludedClassBindingContainer() {
+      run("IncludedClassBindingContainer.kt");
+    }
+
+    @Test
     @TestMetadata("MultibindsOnlyInContainer.kt")
     public void testMultibindsOnlyInContainer() {
       run("MultibindsOnlyInContainer.kt");